REF: Implement BaseMaskedArray class for integer/boolean ExtensionArrays (#30789)
